+++ asset/setup.py
@@ -28,7 +28,7 @@ version = '0.1.1'
 # 'Development Status :: 5 - Production/Stable'
 release_status = 'Development Status :: 3 - Alpha'
 dependencies = [
-    'google-api-core[grpc] >= 1.4.1, < 2.0.0dev',
+    'google-api-core[grpc] >= 1.6.0, < 2.0.0dev',
     'enum34; python_version < "3.4"',
     'grpc-google-iam-v1 >= 0.11.4, < 0.12dev',
 ]
