+++ bigquery_storage/setup.py
@@ -24,7 +24,7 @@ description = 'BigQuery Storage API API client library'
 version = '0.1.0'
 release_status = 'Development Status :: 3 - Alpha'
 dependencies = [
-    'google-api-core[grpc] >= 1.5.1, < 2.0.0dev',
+    'google-api-core[grpc] >= 1.6.0, < 2.0.0dev',
     'enum34; python_version < "3.4"',
 ]
 extras = {
