+++ iot/setup.py
@@ -22,7 +22,7 @@ description = 'Cloud IoT API API client library'
 version = '0.1.0'
 release_status = 'Development Status :: 3 - Alpha'
 dependencies = [
-    'google-api-core[grpc] >= 1.4.1, < 2.0.0dev',
+    'google-api-core[grpc] >= 1.6.0, < 2.0.0dev',
     'grpc-google-iam-v1 >= 0.11.4, < 0.12dev',
     'enum34; python_version < "3.4"',
 ]
