+++ monitoring/setup.py
@@ -29,7 +29,7 @@ version = '0.31.0'
 # 'Development Status :: 5 - Production/Stable'
 release_status = 'Development Status :: 3 - Alpha'
 dependencies = [
-    'google-api-core[grpc] >= 1.4.1, < 2.0.0dev',
+    'google-api-core[grpc] >= 1.6.0, < 2.0.0dev',
 ]
 extras = {
     'pandas': 'pandas >= 0.17.1',
