+++ texttospeech/setup.py
@@ -42,7 +42,7 @@ version = '0.2.0'
 # 'Development Status :: 5 - Production/Stable'
 release_status = 'Development Status :: 3 - Alpha'
 dependencies = [
-    'google-api-core[grpc] >= 1.4.1, < 2.0.0dev',
+    'google-api-core[grpc] >= 1.6.0, < 2.0.0dev',
 ]
 extras = {
 }
